More lib clean-ups

+++ pkg/wfc3/lib/getkeys.c
@@ -1,14 +1,13 @@
-# include <stdio.h>
-# include <stddef.h>
-# include <string.h>
-# include <ctype.h>		/* islower, toupper */
+#include <stdio.h>
+#include <string.h>
+#include <ctype.h>        /* islower, toupper */
 
 #include "hstcal.h"
-# include "hstio.h"
+#include "hstio.h"
 
-# include "wf3.h"
-# include "wf3info.h"
-# include "hstcalerr.h"
+#include "wf3.h"
+#include "wf3info.h"
+#include "hstcalerr.h"
 
 /* This routine gets keyword values from the primary header.
 
@@ -19,22 +18,22 @@
     17-Nov-1998 WJH - Revised to support trailer files
     11-Feb-1999 WJH - Read EXPTIME from Primary header instead of SCI hdr
     17-Apr-2001 HAB - WF3 IR channel mods: changed exptime from scalar to array;
-		      read FILTER keyword instead of FILTER1.
+              read FILTER keyword instead of FILTER1.
      8-May-2001 HAB - Added support for CCD post-flash keywords;
-		      Change UVIS FILTER1 and FILTER2 to just FILTER
-		      (WFC3 UVIS will allow only 1 filter at a time).
+              Change UVIS FILTER1 and FILTER2 to just FILTER
+              (WFC3 UVIS will allow only 1 filter at a time).
     16-Nov-2001 HAB - Updated to use default value of 3 for CCDOFST[A,B,C,D]
     21-Jun-2002 HAB - Updated to read in EXPSTART and EXPEND for computing
-		      darktime, and also read in OBSTYPE.
+              darktime, and also read in OBSTYPE.
     16-Oct-2003 HAB - Updated to use floating-point gain values for WFC3.
     20-Feb-2004 HAB - Eliminated attemp to read BINAXIS keywords from primary
-		      header because for WFC3 they're in the sci extension hdr.
+              header because for WFC3 they're in the sci extension hdr.
     15-Feb-2007 HAB - Updated default gain for IR channel from 2.0 to 2.5.
-                      Added 'subtype' to list of IR keywords loaded. Changed
-                      default sampzero value to 2.911755 sec, to correspond
-		      with new IR timing patterns.
+              Added 'subtype' to list of IR keywords loaded. Changed
+              default sampzero value to 2.911755 sec, to correspond
+              with new IR timing patterns.
     09-Jan-2009 HAB - Eliminated use of default values for FILTER and CCDGAIN
-		      keywords. It will now be an error if they aren't present.
+              keywords. It will now be an error if they aren't present.
     04-Mar-2020 MDD - Read PCTERNOI from the primary header of the raw file to
               determine if it has been set to a non-zero value to be used in
               the CTE reduction.
@@ -43,160 +42,160 @@
 int GetKeys (WF3Info *wf3, Hdr *phdr) {
 
 /* arguments:
-WF3Info *wf3  	io: calibration switches and info
+WF3Info *wf3      io: calibration switches and info
 Hdr *phdr        i: primary header
 */
 
-	extern int status;
+    extern int status;
 
-	int nextend;			/* number of FITS extensions */
-	int i;
-	Bool subarray=0;
+    int nextend;            /* number of FITS extensions */
+    size_t i;
+    Bool subarray=0;
 
-	int GetKeyInt (Hdr *, char *, int, int, int *);
-	int GetKeyFlt (Hdr *, char *, int, float, float *);
-	int GetKeyDbl (Hdr *, char *, int, double, double *);
-	int GetKeyStr (Hdr *, char *, int, char *, char *, int);
-	int GetKeyBool (Hdr *, char *, int, Bool, Bool *);
+    int GetKeyInt (Hdr *, char *, int, int, int *);
+    int GetKeyFlt (Hdr *, char *, int, float, float *);
+    int GetKeyDbl (Hdr *, char *, int, double, double *);
+    int GetKeyStr (Hdr *, char *, int, char *, char *, int);
+    int GetKeyBool (Hdr *, char *, int, Bool, Bool *);
 
-	/* Get generic parameters. */
+    /* Get generic parameters. */
 
-	if (GetKeyStr (phdr, "ROOTNAME", NO_DEFAULT, "", wf3->rootname,SZ_CBUF))
-	    return (status);
+    if (GetKeyStr (phdr, "ROOTNAME", NO_DEFAULT, "", wf3->rootname,SZ_CBUF))
+        return (status);
 
-	if (GetKeyStr (phdr, "APERTURE", USE_DEFAULT, "",wf3->aperture,SZ_CBUF))
-	    return (status);
+    if (GetKeyStr (phdr, "APERTURE", USE_DEFAULT, "",wf3->aperture,SZ_CBUF))
+        return (status);
 
-	if (GetKeyStr (phdr, "OBSTYPE", USE_DEFAULT, "",wf3->obstype,SZ_CBUF))
-	    return (status);
+    if (GetKeyStr (phdr, "OBSTYPE", USE_DEFAULT, "",wf3->obstype,SZ_CBUF))
+        return (status);
 
-	if (GetKeyStr (phdr, "DETECTOR", NO_DEFAULT, "", wf3->det, SZ_CBUF))
-	    return (status);
+    if (GetKeyStr (phdr, "DETECTOR", NO_DEFAULT, "", wf3->det, SZ_CBUF))
+        return (status);
 
-	if (GetKeyBool (phdr, "SUBARRAY", NO_DEFAULT, 0, &subarray))
-		  return (status);
+    if (GetKeyBool (phdr, "SUBARRAY", NO_DEFAULT, 0, &subarray))
+        return (status);
 
-	if (subarray){
-			wf3->subarray=1;
-	}
+    if (subarray) {
+        wf3->subarray=1;
+    }
 
-	if (strcmp (wf3->det, "IR") == 0) {
-	    wf3->detector = IR_DETECTOR;
-	} else if (strcmp (wf3->det, "UVIS") == 0) {
-	    wf3->detector = CCD_DETECTOR;
-	} else {
-	    trlerror("DETECTOR = %s is invalid", wf3->det);
-	    return (status = HEADER_PROBLEM);
-	}
+    if (strcmp (wf3->det, "IR") == 0) {
+        wf3->detector = IR_DETECTOR;
+    } else if (strcmp (wf3->det, "UVIS") == 0) {
+        wf3->detector = CCD_DETECTOR;
+    } else {
+        trlerror("DETECTOR = %s is invalid", wf3->det);
+        return (status = HEADER_PROBLEM);
+    }
 
-	/* Filter or prism/grism name */
-	if (GetKeyStr (phdr, "FILTER", NO_DEFAULT, "", wf3->filter, SZ_CBUF))
-	    return (status);
+    /* Filter or prism/grism name */
+    if (GetKeyStr (phdr, "FILTER", NO_DEFAULT, "", wf3->filter, SZ_CBUF))
+        return (status);
 
-	/* Exposure time */
-	if (GetKeyDbl (phdr, "EXPTIME", NO_DEFAULT, 0., &(wf3->exptime[0])))
-	    return (status);
-	if (wf3->exptime[0] < 0.) {
-	    trlerror("Exposure time is invalid:  %14.6g.", wf3->exptime[0]);
-	    return (status = INVALID_EXPTIME);
-	}
-	if (GetKeyDbl (phdr, "EXPSTART", NO_DEFAULT, 0., &wf3->expstart))
-	    return (status);
-	if (GetKeyDbl (phdr, "EXPEND", NO_DEFAULT, 0., &wf3->expend))
-	    return (status);
+    /* Exposure time */
+    if (GetKeyDbl (phdr, "EXPTIME", NO_DEFAULT, 0., &(wf3->exptime[0])))
+        return (status);
+    if (wf3->exptime[0] < 0.) {
+        trlerror("Exposure time is invalid:  %14.6g.", wf3->exptime[0]);
+        return (status = INVALID_EXPTIME);
+    }
+    if (GetKeyDbl (phdr, "EXPSTART", NO_DEFAULT, 0., &wf3->expstart))
+        return (status);
+    if (GetKeyDbl (phdr, "EXPEND", NO_DEFAULT, 0., &wf3->expend))
+        return (status);
 
-	/* Find out how many extensions there are in this file. */
-	if (GetKeyInt (phdr, "NEXTEND", USE_DEFAULT, EXT_PER_GROUP, &nextend))
-	    return (status);
+    /* Find out how many extensions there are in this file. */
+    if (GetKeyInt (phdr, "NEXTEND", USE_DEFAULT, EXT_PER_GROUP, &nextend))
+        return (status);
 
-	/* Convert number of extensions to number of SingleGroups. */
-	wf3->nimsets = nextend / EXT_PER_GROUP;
-	if (wf3->nimsets < 1) {
-	    trlerror("NEXTEND = %d; must be at least %d.", nextend, EXT_PER_GROUP);
-	    return (status = INVALID_VALUE);
-	}
+    /* Convert number of extensions to number of SingleGroups. */
+    wf3->nimsets = nextend / EXT_PER_GROUP;
+    if (wf3->nimsets < 1) {
+        trlerror("NEXTEND = %d; must be at least %d.", nextend, EXT_PER_GROUP);
+        return (status = INVALID_VALUE);
+    }
 
-	/* Get CCD-specific parameters. */
+    /* Get CCD-specific parameters. */
 
-	if (wf3->detector == CCD_DETECTOR) {
+    if (wf3->detector == CCD_DETECTOR) {
 
-	    if (GetKeyStr (phdr, "CCDAMP", NO_DEFAULT, "", wf3->ccdamp, NAMPS))
-		return (status);
+        if (GetKeyStr (phdr, "CCDAMP", NO_DEFAULT, "", wf3->ccdamp, NAMPS))
+            return (status);
 
-	    for (i=0; i < strlen(wf3->ccdamp) ; i++) {
-			 /* Convert each letter in CCDAMP to upper-case. */
-			 if (islower (wf3->ccdamp[i]))
-			     wf3->ccdamp[i] = toupper (wf3->ccdamp[i]);
+        for (i=0; i < strlen(wf3->ccdamp) ; i++) {
+             /* Convert each letter in CCDAMP to upper-case. */
+             if (islower (wf3->ccdamp[i]))
+                 wf3->ccdamp[i] = toupper (wf3->ccdamp[i]);
 
-			 /* Verify that only the letters 'ABCD' are in the string. */
-			 if (strchr ("ABCD", wf3->ccdamp[i]) == NULL) {
-			     trlerror("CCDAMP = `%s' is invalid.",wf3->ccdamp);
-			     return (status = INVALID_VALUE);
-			 }
-	    }
+             /* Verify that only the letters 'ABCD' are in the string. */
+             if (strchr ("ABCD", wf3->ccdamp[i]) == NULL) {
+                 trlerror("CCDAMP = `%s' is invalid.",wf3->ccdamp);
+                 return (status = INVALID_VALUE);
+             }
+        }
 
-	    if (GetKeyFlt (phdr, "CCDGAIN", NO_DEFAULT, 0., &wf3->ccdgain))
-		return (status);
+        if (GetKeyFlt (phdr, "CCDGAIN", NO_DEFAULT, 0., &wf3->ccdgain))
+            return (status);
 
-	    /* ASSUMPTION: if no CCDOFST values are found, assume they were
-	    ** taken with the default offset setting of 3. This will affect
-	    ** which row is selected from CCDTAB for setting the default value
-	    ** of the bias level in case there is no overscan regions for image.
-	    */
+        /* ASSUMPTION: if no CCDOFST values are found, assume they were
+        ** taken with the default offset setting of 3. This will affect
+        ** which row is selected from CCDTAB for setting the default value
+        ** of the bias level in case there is no overscan regions for image.
+        */
 
-	    if (GetKeyInt (phdr, "CCDOFSTA", USE_DEFAULT, DEFAULT_OFFSET,
-			   &wf3->ccdoffset[0]))
-		return (status);
-	    if (GetKeyInt (phdr, "CCDOFSTB", USE_DEFAULT, DEFAULT_OFFSET,
-			   &wf3->ccdoffset[1]))
-		return (status);
-	    if (GetKeyInt (phdr, "CCDOFSTC", USE_DEFAULT, DEFAULT_OFFSET,
-			   &wf3->ccdoffset[2]))
-		return (status);
-	    if (GetKeyInt (phdr, "CCDOFSTD", USE_DEFAULT, DEFAULT_OFFSET,
-			   &wf3->ccdoffset[3]))
-		return (status);
-	    if (GetKeyFlt (phdr, "FLASHDUR", USE_DEFAULT, 1.0, &wf3->flashdur))
-		return (status);
-	    if (GetKeyStr (phdr, "FLASHSTA", NO_DEFAULT, "", wf3->flashstatus,
-			   SZ_CBUF))
-		return (status);
+        if (GetKeyInt (phdr, "CCDOFSTA", USE_DEFAULT, DEFAULT_OFFSET,
+               &wf3->ccdoffset[0]))
+            return (status);
+        if (GetKeyInt (phdr, "CCDOFSTB", USE_DEFAULT, DEFAULT_OFFSET,
+               &wf3->ccdoffset[1]))
+            return (status);
+        if (GetKeyInt (phdr, "CCDOFSTC", USE_DEFAULT, DEFAULT_OFFSET,
+               &wf3->ccdoffset[2]))
+            return (status);
+        if (GetKeyInt (phdr, "CCDOFSTD", USE_DEFAULT, DEFAULT_OFFSET,
+               &wf3->ccdoffset[3]))
+            return (status);
+        if (GetKeyFlt (phdr, "FLASHDUR", USE_DEFAULT, 1.0, &wf3->flashdur))
+            return (status);
+        if (GetKeyStr (phdr, "FLASHSTA", NO_DEFAULT, "", wf3->flashstatus,
+               SZ_CBUF))
+            return (status);
 
         /* Read the PCTERNOI keyword - If this keyword has a non-zero value in the raw
          * file header, then the CTE algorithm will adopt this value for use in the CTE
          * reduction instead of computing it explicitly.
          */
-	    if (GetKeyFlt (phdr, "PCTERNOI", USE_DEFAULT, 0., &wf3->pcternoi))
-		return (status);
+        if (GetKeyFlt (phdr, "PCTERNOI", USE_DEFAULT, 0., &wf3->pcternoi))
+            return (status);
 
-	} else {
+    } else {
 
-	/* Get IR-specific parameters */
+    /* Get IR-specific parameters */
 
-	    sprintf (wf3->ccdamp, "%s", "ABCD");
+        sprintf (wf3->ccdamp, "%s", "ABCD");
 
-	    if (GetKeyFlt (phdr, "CCDGAIN", NO_DEFAULT, 0., &wf3->ccdgain))
-		return (status);
+        if (GetKeyFlt (phdr, "CCDGAIN", NO_DEFAULT, 0., &wf3->ccdgain))
+            return (status);
 
-	    wf3->nsamp = 0;
-	    if (GetKeyInt (phdr, "NSAMP", NO_DEFAULT, 0, &wf3->nsamp))
-		return (status);
-	    wf3->ngroups = wf3->nsamp;
+        wf3->nsamp = 0;
+        if (GetKeyInt (phdr, "NSAMP", NO_DEFAULT, 0, &wf3->nsamp))
+            return (status);
+        wf3->ngroups = wf3->nsamp;
 
-	    wf3->sampseq[0] = '\0';
-	    if (GetKeyStr (phdr, "SAMP_SEQ", NO_DEFAULT, "", wf3->sampseq,
-			   SZ_CBUF))
-		return (status);
+        wf3->sampseq[0] = '\0';
+        if (GetKeyStr (phdr, "SAMP_SEQ", NO_DEFAULT, "", wf3->sampseq,
+               SZ_CBUF))
+            return (status);
 
-	    wf3->subtype[0] = '\0';
-	    if (GetKeyStr (phdr, "SUBTYPE", NO_DEFAULT, "", wf3->subtype,
-			   SZ_CBUF))
-		return (status);
+        wf3->subtype[0] = '\0';
+        if (GetKeyStr (phdr, "SUBTYPE", NO_DEFAULT, "", wf3->subtype,
+               SZ_CBUF))
+            return (status);
 
-	    wf3->sampzero = 0.0;
-	    if (GetKeyDbl (phdr, "SAMPZERO", USE_DEFAULT, 2.911755, &wf3->sampzero))
-		return (status);
-	}
+        wf3->sampzero = 0.0;
+        if (GetKeyDbl (phdr, "SAMPZERO", USE_DEFAULT, 2.911755, &wf3->sampzero))
+            return (status);
+    }
 
-	return (status);
+    return (status);
 }
